Refactor NCActivity & NCActivityCell
- remove unnecessary, old code & comments
- remove unnecessary storyboard views
- factor out TableViewCell into new file
